# backend/api-service/src/main/java/com/tripflow/dto/itinerary/ExtendedItineraryResponseDTO.java
package com.tripflow.dto.itinerary;

import com.fasterxml.jackson.annotation.JsonUnwrapped;

public record ExtendedItineraryResponseDTO(
    @JsonUnwrapped
    ExtendedItineraryDTO itinerary,
    PermissionsDTO permissions
) {}