# docs/api/api-docs.yaml
openapi: 3.1.0
info:
  title: TripFlow API
  description: "API documentation for TripFlow, the travel planning application."
  version: v1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- auth_token: []
...
paths:
...
  /api/v1/itineraries/{itineraryId}/collaborators/{username}:
    put:
      tags:
      - Itinerary Collaboration
      summary: Update Collaborator Role
      description: Updates a collaborator's role. Only owner can do this.
      operationId: updateCollaboratorRole
      parameters:
      - name: itineraryId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: username
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCollaboratorRequest"
        required: true
      responses:
        "200":
          description: Collaborator updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CollaboratorDTO"
        "400":
          description: Invalid request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CollaboratorDTO"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CollaboratorDTO"
        "404":
          description: Itinerary or collaborator not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CollaboratorDTO"
      security:
      - auth_token: []
    delete:
      tags:
      - Itinerary Collaboration
      summary: Remove Collaborator
      description: "Removes a collaborator. Owner can remove anyone, collaborators\
        \ can remove themselves."
      operationId: removeCollaborator
      parameters:
      - name: itineraryId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Collaborator removed
        "403":
          description: Forbidden
        "404":
          description: Itinerary or collaborator not found
      security:
      - auth_token: []
  /api/v1/itineraries/{itineraryId}/collaborators/{username}/accept:
    put:
      tags:
      - Itinerary Collaboration
      summary: Accept Invitation
      description: Accepts a pending invitation to collaborate on an itinerary.
      operationId: acceptInvitation
      parameters:
      - name: itineraryId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Invitation accepted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CollaboratorDTO"
        "400":
          description: Invitation is not pending
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CollaboratorDTO"
        "404":
          description: Invitation not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CollaboratorDTO"
      security:
      - auth_token: []
...
  /api/v1/itineraries/{itineraryId}/collaborators:
    get:
      tags:
      - Itinerary Collaboration
      summary: Get Collaborators
      description: Gets all collaborators of an itinerary.
      operationId: getCollaborators
      parameters:
      - name: itineraryId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Collaborators found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollaboratorDTO"
        "404":
          description: Itinerary not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollaboratorDTO"
      security:
      - auth_token: []
    post:
      tags:
      - Itinerary Collaboration
      summary: Send Invitation
      description: Sends an invitation to a user to collaborate on an itinerary. Only
        owner can do this.
      operationId: sendInvitation
      parameters:
      - name: itineraryId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCollaboratorRequest"
        required: true
      responses:
        "201":
          description: Invitation sent
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CollaboratorDTO"
        "400":
          description: Invalid request or user is owner
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CollaboratorDTO"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CollaboratorDTO"
        "404":
          description: Itinerary or user not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CollaboratorDTO"
        "409":
          description: User already has a pending invitation or is a collaborator
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CollaboratorDTO"
      security:
      - auth_token: []
  /api/v1/itineraries:
    get:
      tags:
      - Itinerary Management
      summary: Get All Itineraries
      description: Retrieves a paginated list of itineraries for the authenticated
        user.
      operationId: getAllItineraries
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      - name: search
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Itineraries retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaginatedDTOItineraryResponseDTO"
        "401":
          description: Unauthorized access
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaginatedDTOItineraryResponseDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaginatedDTOItineraryResponseDTO"
      security:
      - auth_token: []
    post:
      tags:
      - Itinerary Management
      summary: Create Itinerary
      description: Creates a new itinerary for the authenticated user.
      operationId: createItinerary
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtendedItineraryDTO"
        required: true
      responses:
        "201":
          description: Itinerary created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExtendedItineraryResponseDTO"
        "400":
          description: Invalid itinerary data provided
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExtendedItineraryResponseDTO"
        "401":
          description: Unauthorized access
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExtendedItineraryResponseDTO"
      security:
      - auth_token: []
  /api/v1/itineraries/:
    get:
      tags:
      - Itinerary Management
      summary: Get All Itineraries
      description: Retrieves a paginated list of itineraries for the authenticated
        user.
      operationId: getAllItineraries_1
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      - name: search
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Itineraries retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaginatedDTOItineraryResponseDTO"
        "401":
          description: Unauthorized access
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaginatedDTOItineraryResponseDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaginatedDTOItineraryResponseDTO"
      security:
      - auth_token: []
    post:
      tags:
      - Itinerary Management
      summary: Create Itinerary
      description: Creates a new itinerary for the authenticated user.
      operationId: createItinerary_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtendedItineraryDTO"
        required: true
      responses:
        "201":
          description: Itinerary created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExtendedItineraryResponseDTO"
        "400":
          description: Invalid itinerary data provided
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExtendedItineraryResponseDTO"
        "401":
          description: Unauthorized access
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExtendedItineraryResponseDTO"
      security:
      - auth_token: []
...
  /api/v1/itineraries/{itineraryId}/collaborators/{username}/decline:
    delete:
      tags:
      - Itinerary Collaboration
      summary: Decline Invitation
      description: Declines a pending invitation to collaborate on an itinerary.
      operationId: declineInvitation
      parameters:
      - name: itineraryId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Invitation declined
        "400":
          description: Invitation is already accepted or invalid
        "404":
          description: Invitation not found
      security:
      - auth_token: []
components:
  schemas:
...
    PublicUserDTO:
      type: object
      properties:
        username:
          type: string
        name:
          type: string
        description:
          type: string
        location:
          type: string
        notificationsAllowed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        role:
          type: string
          enum:
          - USER
          - ADMIN
        plan:
          type: string
          enum:
          - FREE
          - PRO
          - PREMIUM
    UpdateCollaboratorRequest:
      type: object
      properties:
        role:
          type: string
          enum:
          - EDITOR
          - VIEWER
          - OWNER
      required:
      - role
    CollaboratorDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: "#/components/schemas/PublicUserDTO"
        fromUser:
          type: string
        role:
          type: string
          enum:
          - EDITOR
          - VIEWER
          - OWNER
        status:
          type: string
          enum:
          - PENDING
          - ACCEPTED
        invitedAt:
          type: string
          format: date-time
        acceptedAt:
          type: string
          format: date-time
        itineraryId:
          type: integer
          format: int64
        itineraryTitle:
          type: string
    ActivityDTO:
      type: object
      properties:
        activity:
          type: string
        details:
          type: string
        location:
          $ref: "#/components/schemas/LocationDTO"
        time:
          type: string
        duration:
          type: string
    CoordinatesDTO:
      type: object
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    ExtendedItineraryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          maxLength: 200
          minLength: 0
        place:
          type: string
          maxLength: 200
          minLength: 0
        people:
          type: integer
          format: int32
        budget:
          type: number
          format: double
        date:
          type: string
          maxLength: 50
          minLength: 0
        tags:
          type: array
          items:
            type: string
          maxItems: 30
          minItems: 0
        updatedCount:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - DRAFT
          - PLANNED
          - ONGOING
          - COMPLETED
        days:
          type: array
          items:
            $ref: "#/components/schemas/ItineraryDayDTO"
        countDays:
          type: integer
          format: int32
        coverImage:
          $ref: "#/components/schemas/ExternalImageDTO"
      required:
      - title
    ExternalImageDTO:
      type: object
      properties:
        imageUrl:
          type: string
        altDescription:
          type: string
        authorUsername:
          type: string
    ItineraryDayDTO:
      type: object
      properties:
        day:
          type: integer
          format: int32
        activities:
          type: array
          items:
            $ref: "#/components/schemas/ActivityDTO"
    LocationDTO:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        coordinates:
          $ref: "#/components/schemas/CoordinatesDTO"
    ExtendedItineraryResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          maxLength: 200
          minLength: 0
        place:
          type: string
          maxLength: 200
          minLength: 0
        people:
          type: integer
          format: int32
        budget:
          type: number
          format: double
        date:
          type: string
          maxLength: 50
          minLength: 0
        tags:
          type: array
          items:
            type: string
          maxItems: 30
          minItems: 0
        updatedCount:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - DRAFT
          - PLANNED
          - ONGOING
          - COMPLETED
        days:
          type: array
          items:
            $ref: "#/components/schemas/ItineraryDayDTO"
        countDays:
          type: integer
          format: int32
        coverImage:
          $ref: "#/components/schemas/ExternalImageDTO"
        permissions:
          $ref: "#/components/schemas/PermissionsDTO"
      required:
      - title
    PermissionsDTO:
      type: object
      properties:
        view:
          type: boolean
        edit:
          type: boolean
        delete:
          type: boolean
...
    AddCollaboratorRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        role:
          type: string
          enum:
          - EDITOR
          - VIEWER
          - OWNER
      required:
      - role
      - username
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    ItineraryResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        place:
          type: string
        people:
          type: integer
          format: int32
        budget:
          type: number
          format: double
        date:
          type: string
        countDays:
          type: integer
          format: int32
        tags:
          type: array
          items:
            type: string
        updatedCount:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - DRAFT
          - PLANNED
          - ONGOING
          - COMPLETED
        coverImage:
          $ref: "#/components/schemas/ExternalImageDTO"
        permissions:
          $ref: "#/components/schemas/PermissionsDTO"
    PaginatedDTOItineraryResponseDTO:
      type: object
      properties:
        page:
          type: array
          items:
            $ref: "#/components/schemas/ItineraryResponseDTO"
        currentPage:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
        totalItems:
          type: integer
          format: int64
        itemsPerPage:
          type: integer
          format: int64
        isLastPage:
          type: boolean
...
  securitySchemes:
    auth_token:
      type: apiKey
      name: auth_token
      in: cookie